// Frontend/src/core/models/SignUp.ts
export interface SignUp {
    name: string;
    email: string;
    password: string;
    confirmPassword: string;
    number: string;
    city: string;
    zipcode: string;
    fullAddress: string;
    description: string;
    website: string;
    linkedin: string;
}
export interface SignUpDev {
    name: string;
    email: string;
    password: string;
    confirmPassword: string;

    number: string;
    city: string;
    zipcode: string;
    fullAddress: string;
    currentPosition: string;
    linkedin: string;
    github: string;
}